+++ sdk/objc/api/peerconnection/RTCPeerConnectionFactory.mm
@@ -25,6 +25,7 @@
 #import "RTCMediaStream+Private.h"
 #import "RTCPeerConnection+Private.h"
 #import "RTCVideoSource+Private.h"
+#import "RTCAudioProcessingState+Private.h"
 #import "RTCVideoTrack+Private.h"
 #import "RTCRtpReceiver+Private.h"
 #import "RTCRtpCapabilities+Private.h"
@@ -422,6 +423,11 @@ static webrtc::Environment CreateDefaultEnvironment() {
 
   webrtc::RtpCapabilities rtpCapabilities = _nativeFactory->GetRtpReceiverCapabilities(mediaType);
   return [[RTC_OBJC_TYPE(RTCRtpCapabilities) alloc] initWithNativeRtpCapabilities:rtpCapabilities];
+}
+
+- (RTC_OBJC_TYPE(RTCAudioProcessingRuntimeState))audioProcessingRuntimeState {
+  return webrtc::objc::AudioProcessingRuntimeStateToObjC(
+      _nativeFactory->GetAudioProcessingRuntimeState());
 }
 
 - (RTC_OBJC_TYPE(RTCAudioSource) *)audioSourceWithConstraints:
